requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
pandas==1.5.3
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
requests==2.31.0
python-Levenshtein==0.21.1
openpyxl==3.1.2
matplotlib==3.7.2
werkzeug==2.3.7